Extract Most
